Rename Switch Template to Use Template
